requirements.txt
amqp==5.3.1
asgiref==3.8.1
bcrypt==4.3.0
billiard==4.2.1
celery==5.5.2
certifi==2024.6.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.2.0
cycler==0.12.1
distlib==0.3.8
dj-rest-auth==7.0.1
Django==4.2.20
django-allauth==65.7.0
django-cors-headers==4.7.0
django-environ==0.11.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
filelock==3.15.4
fonttools==4.50.0
fsspec==2025.2.0
imageio==2.34.1
importlib_resources==6.4.0
Jinja2==3.1.5
joblib==1.3.2
kiwisolver==1.4.5
kombu==5.5.3
MarkupSafe==3.0.2
matplotlib==3.8.4
mpmath==1.3.0
mysqlclient==2.2.4
networkx==3.2.1
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pillow==10.2.0
pipenv==2024.0.1
platformdirs==4.2.2
prompt_toolkit==3.0.51
PyJWT==2.9.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2024.1
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
sqlparse==0.5.3
sympy==1.13.1
threadpoolctl==3.4.0
torch==2.6.0
typing_extensions==4.12.2
tzdata==2025.2
ucimlrepo==0.0.7
vine==5.1.0
virtualenv==20.26.3
wcwidth==0.2.13
zipp==3.18.1
